+++ presets/custom_weight_paint_colors/blue_yellow.py
@@ -1,5 +1,0 @@
-import bpy
-prefs = bpy.context.preferences
-
-prefs.view.use_weight_color_range = True
-prefs.view.weight_color_range.color_mode = 'RGB'