+++ presets/custom_weight_paint_colors/punk_constant.py
@@ -16,10 +16,10 @@ for elm in range(len(colRangeElm)):
 #Reset first to 0,0
 colRangeElm[0].position=float(0.0)
 
-posList= [0.0, 0.009999999776482582, 0.7328363656997681, 1.0]
-colList= [(0.31110239028930664, 0.31110239028930664, 0.31110239028930664, 1.0), (0.004202490672469139, 0.5928969383239746, 0.11618726700544357, 1.0), (0.7453177571296692, 1.0, 0.004896974191069603, 1.0), (1.0, 1.0, 1.0, 1.0)]
+posList= [0.0, 0.009999999776482582, 0.5, 0.75, 1.0]
+colList= [(0.31110239028930664, 0.31110239028930664, 0.31110239028930664, 1.0), (0.006538354326039553, 0.9655017256736755, 1.0, 1.0), (1.0, 0.0031068515963852406, 0.8270024657249451, 1.0), (1.0, 0.8930725455284119, 0.0038679656572639942, 1.0), (1.0, 1.0, 1.0, 1.0)]
 
-for i in range(0,len(colList)):
+for i in range(len(colList)):
     # Update first and last > dont add new
     if i == 0 or i == (len(colList)):
         colRangeElm[i].color = colList[i]
@@ -32,4 +32,4 @@ for i in range(0,len(colList)):
 
 prefs.view.use_weight_color_range = True
 prefs.view.weight_color_range.color_mode = 'RGB'
-prefs.view.weight_color_range.interpolation = 'LINEAR'
+prefs.view.weight_color_range.interpolation = 'CONSTANT'