+++ CHANGELOG.md
@@ -1,26 +1,33 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.4] - 2021-08-02
+### Fixed
+- Issue where 1 additional colorstop was added on saving a preset
+
+### Added
+- Copy couple default presets over
+
 
 ## [0.0.3] - 2021-07-28
 ### Added
 - The menu is now also available in options, allows for fsater sqitch when weightpainting
 
-
 ## [0.0.2] - 2021-07-26
 ### Fixed
 - Save works now  (old test code saved in wrong path)
-
 
 ## [0.0.1] - 2021-07-21
 ### Added
 - Initial release
 
 ## Notes
+
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!--### Official Rigify Info-->
 
+[0.0.4]:https://github.com/schroef/Save-Custom-Weight-Paint-Presets/releases/tag/v0.0.4
 [0.0.3]:https://github.com/schroef/Save-Custom-Weight-Paint-Presets/releases/tag/v0.0.3
 [0.0.2]:https://github.com/schroef/Save-Custom-Weight-Paint-Presets/releases/tag/v0.0.2
 [0.0.1]:https://github.com/schroef/Save-Custom-Weight-Paint-Presets/releases/tag/v0.0.1